Se bajan cambios

+++ Web/digital-booking-c3-e-1/src/components/products/Crud/Crud.jsx
@@ -6,7 +6,7 @@ import Pagination from "../../resources/pagination/Pagination";
 import { Link } from "react-router-dom";
 import { ProductsContext } from "../../../context/ProductsContext";
 import ProductsService from "../../../shared/services/ProductsService";
-
+import Swal from 'sweetalert2';
 import { useNavigate } from 'react-router-dom';
 
 const Crud = () => {
@@ -15,12 +15,22 @@ const Crud = () => {
   const data = useContext(ProductsContext)
   const [currentProducts, setCurrentProducts] = useState([]);
   const [currentPage, setCurrentPage] = useState(1);
+  const [ reload, setReload ] = useState(false)
 
   const pageLimit = 5;
-  
+
   useEffect(() => {
-    setProducts(data.products);
-  }, [data]);
+    const fetchProducts = async () => {
+      try {
+        const prodsActual = await ProductsService.getAll()
+        setProducts(prodsActual)
+      }
+      catch {
+        setProducts(data.products);
+      }
+    }
+    fetchProducts()
+  }, [reload]);
 
   useEffect(() => {
     onPageChanged();
@@ -31,37 +41,59 @@ const Crud = () => {
     setCurrentProducts(products.slice(offset, offset + pageLimit));
   };
 
-// usecallback para memorizar y asegurarnos de que no se creara una nueva instancia en cada renderizado
-  const handleDelete = useCallback(async (productId) => {
-    try {
-      const res = await ProductsService.deleteByID(productId)
-      console.log(res);
-      if(res.status == 200){
-        const updatedProducts = products.filter(
-          (product) => product.id !== productId
-          );
-          setProducts(updatedProducts);
-          console.log("Producto eliminado con ID:", productId);
+  // usecallback para memorizar y asegurarnos de que no se creara una nueva instancia en cada renderizado
+  const handleDelete = (productId) => {
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#a6cf7e',
+      cancelButtonColor: '#fd7053',
+      cancelButtonText: 'No',
+      confirmButtonText: 'Yes, delete it!'
+    }).then( async(result) => {
+      if (result.isConfirmed) {
+        try {
+          const res = await ProductsService.deleteByID(productId)
+          console.log(res);
+          if (res.status == 200) {
+            const updatedProducts = products.filter(
+              (product) => product.id !== productId
+            );
+            setProducts(updatedProducts);
+            setReload(!reload)
+            console.log("Producto eliminado con ID:", productId);
+            Swal.fire(
+              'Deleted!',
+              'Your file has been deleted.',
+              'success'
+            )
+          }
+        } catch (error) {
+          console.log("Error al eliminar el producto", error);
         }
-    } catch (error) {
-      console.log("Error al eliminar el producto", error);
-    }
-  }, [products]);
+      }
+      else {
+        Swal.close();
+      }
+    });
+  }
 
-const handleEdit = useCallback((product) => {
-  console.log("Editando producto:", product);
-  console.log("Editando producto con ID:", product.id);
+  const handleEdit = useCallback((product) => {
+    console.log("Editando producto:", product);
+    console.log("Editando producto con ID:", product.id);
 
-  navigate('product/edit', {state:{product:product}});
-}, [navigate]);
+    navigate('product/edit', { state: { product: product } });
+  }, [navigate]);
 
   return (
     <>
       <section className={styles["container"]} >
         <div className={styles["button-container"]}>
-        <Link to="product/add">
-          <ButtonPrimary>Agregar producto</ButtonPrimary>
-        </Link>
+          <Link to="product/add">
+            <ButtonPrimary>Agregar producto</ButtonPrimary>
+          </Link>
         </div>
         <table className={styles["table"]}>
           <thead>
@@ -89,13 +121,13 @@ const handleEdit = useCallback((product) => {
             ))}
           </tbody>
         </table>
-            <Pagination
-        onPageChanged={onPageChanged}
-        limit={pageLimit}
-        total={products.length}
-        currentPage={currentPage}
-        setCurrentPage={setCurrentPage}
-      />
+        <Pagination
+          onPageChanged={onPageChanged}
+          limit={pageLimit}
+          total={products.length}
+          currentPage={currentPage}
+          setCurrentPage={setCurrentPage}
+        />
       </section>
     </>
   );
